+++ src/components/Dropdown/__stories__/dropdown.stories.scss
@@ -1,3 +1,7 @@
 .dropdown-story {
   min-width: 250px;
 }
+
+.dropdown-limited-width {
+  width: 250px;
+}
