+++ src/components/Dropdown/components/ValueContainer/ValueContainer.module.scss
@@ -1,5 +1,5 @@
-@import "../../styles/themes.scss";
-@import "../../styles/typography.scss";
+@import "../../../../styles/themes.scss";
+@import "../../../../styles/typography.scss";
 
 .multiselect-chip {
   margin: 5px 4px;
