--- src/components/MultiSelect/__stories__/multiSelect.stories.scss
@@ -1,3 +1,0 @@
-.multiselect-story {
-  width: 250px;
-}
